Update background processes
