+++ app/background/__init__.py
@@ -1,0 +1,1 @@
+from app.background.daemon import Daemon  # noqa
