+++ app/config.py
@@ -4,15 +4,12 @@ import os
 class Config:
     """Configures the application."""
 
-    STATICE_REQUEST_INTERVAL = os.environ.get("STATICE_REQUEST_INTERVAL") or 10
+    STATICE_DAEMON_INTERVAL = os.environ.get("STATICE_DAEMON_INTERVAL") or 10
     STATICE_REQUEST_TIMEOUT = os.environ.get("STATICE_REQUEST_TIMEOUT") or 3
-
-    STATICE_PURGE_INTERVAL = os.environ.get("STATICE_PURGE_INTERVAL") or 10
-    STATICE_PURGE_AGE = os.environ.get("STATICE_PURGE_AGE") or 600
 
     SECRET_KEY = os.environ.get("SECRET_KEY") or "change-me"
 
-    POSTGRES_URL = os.environ.get("POSTGRES_URL") or "statice_db:5432"
+    POSTGRES_URL = os.environ.get("POSTGRES_URL") or "localhost:5432"
     POSTGRES_DB = os.environ.get("POSTGRES_DB") or "postgres"
     POSTGRES_USER = os.environ.get("POSTGRES_USER") or "postgres"
     POSTGRES_PASSWORD = os.environ.get("POSTGRES_PASSWORD") or "postgres"
@@ -22,5 +19,5 @@ class Config:
     )
     SQLALCHEMY_TRACK_MODIFICATIONS = False
 
-    RQ_REDIS_URL = os.environ.get("REDIS_URL") or "redis://statice_redis:6379/0"
+    RQ_REDIS_URL = os.environ.get("REDIS_URL") or "redis://localhost:6379/0"
     RQ_QUEUES = ("default",)
