--- app/tasks/__init__.py
@@ -1,2 +1,0 @@
-from app.tasks.janitor import Janitor  # noqa
-from app.tasks.inquisitor import Inquisitor  # noqa
