+++ pyproject.toml
@@ -17,6 +17,7 @@ psycopg2 = "^2.8.4"
 flask-rq2 = "^18.3"
 arrow = "^0.15.5"
 bootstrap-flask = "^1.2.0"
+rq_scheduler = "^0.9.1"
 
 [tool.poetry.dev-dependencies]
 black = {version = "^19.10b0", allow-prereleases = true}
